requirements.txt
streamlit
euriai
pinecone
python-dotenv
langchain
